Adding swagger UI. Adding DefaultExceptionHandle and ValidationHandler

+++ src/main/java/lk/devfactory/controllers/RepositoryController.java
@@ -1,13 +1,16 @@
 package lk.devfactory.controllers;
 
+import java.util.List;
 import java.util.stream.Collectors;
 
+import javax.servlet.http.HttpServletResponse;
 import javax.ws.rs.core.Response.Status;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.util.CollectionUtils;
 
 import com.github.javafaker.Faker;
 
@@ -32,7 +35,7 @@ public class RepositoryController {
     @Autowired
     RepositoryDO repository;
 
-    //TODO: Handle 405
+    //TODO: Handle 405 invalid input git url
     public ResponseContext addRepo(RequestContext request, Repository body) {
     	UUID uuid = UUIDGenerator.get();
     	log.info("Id for repo : "+ body + " is "+uuid);
@@ -43,23 +46,30 @@ public class RepositoryController {
                 .entity(body);
     }
     
-    //TODO: Handle 404
     public ResponseContext findRepositories(RequestContext request) {
     	log.debug("Find all repositories");
-        return new ResponseContext()
+        final List<Repository> repositories = repository.findAll().collect(Collectors.toList());
+        if (CollectionUtils.isEmpty(repositories)) {
+    		log.info("Unable to find any repository");
+    		ApiError error = new ApiError();
+    		error.setCode(HttpServletResponse.SC_NOT_FOUND);
+    		error.setMessage("Unable to find any repository");
+    		throw new ApiException(error);
+        }
+		return new ResponseContext()
                 .status(Status.OK)
-                .entity(repository.findAll().collect(Collectors.toList()));
+                .entity(repositories);
     }
 
-    //TODO: Handle 404 400
+    //TODO: 400
     public ResponseContext getRepoDeadCodeByID(RequestContext requestContext, String id)
     {
     	log.info("Id of the repo to find is "+ id);
-    	Repository repositoryObj = repository.findById(UUIDGenerator.get(id));
+    	final Repository repositoryObj = repository.findById(UUIDGenerator.get(id));
     	if (repositoryObj == null) {
     		log.info("Unable to find repo with "+ id);
     		ApiError error = new ApiError();
-    		error.setCode(404);
+    		error.setCode(HttpServletResponse.SC_NOT_FOUND);
     		error.setMessage("Unable to find repository with id "+id);
     		throw new ApiException(error);
     	}
@@ -67,10 +77,20 @@ public class RepositoryController {
                                     .entity(repositoryObj);
     }
     
-  //TODO: Handle 404 400
+  //TODO: Handle 400
     public ResponseContext removeRepo(RequestContext requestContext, String id)
     {
     	log.info("Id of the repo to remove "+" is "+ id);
+    	
+    	final Repository repositoryObj = repository.findById(UUIDGenerator.get(id));
+    	if (repositoryObj == null) {
+    		log.info("Unable to find repo with "+ id);
+    		ApiError error = new ApiError();
+    		error.setCode(HttpServletResponse.SC_NOT_FOUND);
+    		error.setMessage("Unable to find repository with id "+id);
+    		throw new ApiException(error);
+    	}
+    	
     	repository.remove(UUIDGenerator.get(id));
     	log.info("Id of the repo removed "+" is "+ id);
         return new ResponseContext().status(Status.OK);
